remove `EigenConfig::default()`
