+++ core/lib/config/src/configs/da_client/eigen.rs
@@ -1,13 +1,5 @@
-use std::str::FromStr;
-
 use serde::Deserialize;
-use zksync_basic_types::{secrets::PrivateKey, url::SensitiveUrl, Address, H160};
-
-/// Default address of the EigenDA service manager contract deployed on Holesky.
-const DEFAULT_EIGENDA_SVC_MANAGER_ADDRESS: H160 = H160([
-    0xd4, 0xa7, 0xe1, 0xbd, 0x80, 0x15, 0x05, 0x72, 0x93, 0xf0, 0xd0, 0xa5, 0x57, 0x08, 0x8c, 0x28,
-    0x69, 0x42, 0xe8, 0x4b,
-]);
+use zksync_basic_types::{secrets::PrivateKey, url::SensitiveUrl, Address};
 
 /// Configuration for the EigenDA remote disperser client.
 #[derive(Clone, Debug, PartialEq, Deserialize)]
@@ -31,22 +23,6 @@ pub struct EigenConfig {
     pub g1_url: String,
     /// Url to the file containing the G2 point used for KZG
     pub g2_url: String,
-}
-
-impl Default for EigenConfig {
-    fn default() -> Self {
-        Self {
-            disperser_rpc: "https://disperser-holesky.eigenda.xyz:443".to_string(),
-            settlement_layer_confirmation_depth: 0,
-            eigenda_eth_rpc: Some(SensitiveUrl::from_str("https://ethereum-holesky-rpc.publicnode.com").unwrap()), // Safe to unwrap, never fails
-            eigenda_svc_manager_address: DEFAULT_EIGENDA_SVC_MANAGER_ADDRESS,
-            wait_for_finalization: false,
-            authenticated: false,
-            points_dir: None,
-            g1_url: "https://github.com/Layr-Labs/eigenda-proxy/raw/2fd70b99ef5bf137d7bbca3461cf9e1f2c899451/resources/g1.point".to_string(),
-            g2_url: "https://github.com/Layr-Labs/eigenda-proxy/raw/2fd70b99ef5bf137d7bbca3461cf9e1f2c899451/resources/g2.point.powerOf2".to_string(),
-        }
-    }
 }
 
 #[derive(Clone, Debug, PartialEq)]
