+++ core/node/da_clients/src/eigen/verifier/tests.rs
@@ -107,10 +107,29 @@ fn create_remote_query_client() -> Box<DynClient<L1>> {
     Box::new(query_client) as Box<DynClient<L1>>
 }
 
+const DEFAULT_EIGENDA_SVC_MANAGER_ADDRESS: H160 = H160([
+    0xd4, 0xa7, 0xe1, 0xbd, 0x80, 0x15, 0x05, 0x72, 0x93, 0xf0, 0xd0, 0xa5, 0x57, 0x08, 0x8c, 0x28,
+    0x69, 0x42, 0xe8, 0x4b,
+]);
+
+fn test_eigen_config() -> EigenConfig {
+    EigenConfig {
+            disperser_rpc: "https://disperser-holesky.eigenda.xyz:443".to_string(),
+            settlement_layer_confirmation_depth: 0,
+            eigenda_eth_rpc: Some(SensitiveUrl::from_str("https://ethereum-holesky-rpc.publicnode.com").unwrap()), // Safe to unwrap, never fails
+            eigenda_svc_manager_address: DEFAULT_EIGENDA_SVC_MANAGER_ADDRESS,
+            wait_for_finalization: false,
+            authenticated: false,
+            points_dir: None,
+            g1_url: "https://github.com/Layr-Labs/eigenda-proxy/raw/2fd70b99ef5bf137d7bbca3461cf9e1f2c899451/resources/g1.point".to_string(),
+            g2_url: "https://github.com/Layr-Labs/eigenda-proxy/raw/2fd70b99ef5bf137d7bbca3461cf9e1f2c899451/resources/g2.point.powerOf2".to_string(),
+        }
+}
+
 #[ignore = "depends on external RPC"]
 #[tokio::test]
 async fn test_verify_commitment() {
-    let cfg = EigenConfig::default();
+    let cfg = test_eigen_config();
     let query_client = create_remote_query_client();
     let verifier = Verifier::new(cfg, Arc::new(query_client)).await.unwrap();
     let commitment = G1Commitment {
@@ -132,7 +151,7 @@ async fn test_verify_commitment() {
 /// To test actual behaviour of the verifier, run the test above
 #[tokio::test]
 async fn test_verify_commitment_mocked() {
-    let cfg = EigenConfig::default();
+    let cfg = test_eigen_config();
     let query_client = MockVerifierClient::new(HashMap::new());
     let verifier = Verifier::new(cfg, Arc::new(query_client)).await.unwrap();
     let commitment = G1Commitment {
@@ -153,7 +172,7 @@ async fn test_verify_commitment_mocked() {
 #[ignore = "depends on external RPC"]
 #[tokio::test]
 async fn test_verify_merkle_proof() {
-    let cfg = EigenConfig::default();
+    let cfg = test_eigen_config();
     let query_client = create_remote_query_client();
     let verifier = Verifier::new(cfg, Arc::new(query_client)).await.unwrap();
     let cert = BlobInfo {
@@ -234,7 +253,7 @@ async fn test_verify_merkle_proof() {
 /// To test actual behaviour of the verifier, run the test above
 #[tokio::test]
 async fn test_verify_merkle_proof_mocked() {
-    let cfg = EigenConfig::default();
+    let cfg = test_eigen_config();
     let query_client = MockVerifierClient::new(HashMap::new());
     let verifier = Verifier::new(cfg, Arc::new(query_client)).await.unwrap();
     let cert = BlobInfo {
@@ -314,7 +333,7 @@ async fn test_verify_merkle_proof_mocked() {
 #[ignore = "depends on external RPC"]
 #[tokio::test]
 async fn test_hash_blob_header() {
-    let cfg = EigenConfig::default();
+    let cfg = test_eigen_config();
     let query_client = create_remote_query_client();
     let verifier = Verifier::new(cfg, Arc::new(query_client)).await.unwrap();
     let blob_header = BlobHeader {
@@ -353,7 +372,7 @@ async fn test_hash_blob_header() {
 /// To test actual behaviour of the verifier, run the test above
 #[tokio::test]
 async fn test_hash_blob_header_mocked() {
-    let cfg = EigenConfig::default();
+    let cfg = test_eigen_config();
     let query_client = MockVerifierClient::new(HashMap::new());
     let verifier = Verifier::new(cfg, Arc::new(query_client)).await.unwrap();
     let blob_header = BlobHeader {
@@ -391,7 +410,7 @@ async fn test_hash_blob_header_mocked() {
 #[ignore = "depends on external RPC"]
 #[tokio::test]
 async fn test_inclusion_proof() {
-    let cfg = EigenConfig::default();
+    let cfg = test_eigen_config();
     let query_client = create_remote_query_client();
     let verifier = Verifier::new(cfg, Arc::new(query_client)).await.unwrap();
     let proof = hex::decode("c455c1ea0e725d7ea3e5f29e9f48be8fc2787bb0a914d5a86710ba302c166ac4f626d76f67f1055bb960a514fb8923af2078fd84085d712655b58a19612e8cd15c3e4ac1cef57acde3438dbcf63f47c9fefe1221344c4d5c1a4943dd0d1803091ca81a270909dc0e146841441c9bd0e08e69ce6168181a3e4060ffacf3627480bec6abdd8d7bb92b49d33f180c42f49e041752aaded9c403db3a17b85e48a11e9ea9a08763f7f383dab6d25236f1b77c12b4c49c5cdbcbea32554a604e3f1d2f466851cb43fe73617b3d01e665e4c019bf930f92dea7394c25ed6a1e200d051fb0c30a2193c459f1cfef00bf1ba6656510d16725a4d1dc031cb759dbc90bab427b0f60ddc6764681924dda848824605a4f08b7f526fe6bd4572458c94e83fbf2150f2eeb28d3011ec921996dc3e69efa52d5fcf3182b20b56b5857a926aa66605808079b4d52c0c0cfe06923fa92e65eeca2c3e6126108e8c1babf5ac522f4d7").unwrap();
@@ -412,7 +431,7 @@ async fn test_inclusion_proof() {
 /// To test actual behaviour of the verifier, run the test above
 #[tokio::test]
 async fn test_inclusion_proof_mocked() {
-    let cfg = EigenConfig::default();
+    let cfg = test_eigen_config();
     let query_client = MockVerifierClient::new(HashMap::new());
     let verifier = Verifier::new(cfg, Arc::new(query_client)).await.unwrap();
     let proof = hex::decode("c455c1ea0e725d7ea3e5f29e9f48be8fc2787bb0a914d5a86710ba302c166ac4f626d76f67f1055bb960a514fb8923af2078fd84085d712655b58a19612e8cd15c3e4ac1cef57acde3438dbcf63f47c9fefe1221344c4d5c1a4943dd0d1803091ca81a270909dc0e146841441c9bd0e08e69ce6168181a3e4060ffacf3627480bec6abdd8d7bb92b49d33f180c42f49e041752aaded9c403db3a17b85e48a11e9ea9a08763f7f383dab6d25236f1b77c12b4c49c5cdbcbea32554a604e3f1d2f466851cb43fe73617b3d01e665e4c019bf930f92dea7394c25ed6a1e200d051fb0c30a2193c459f1cfef00bf1ba6656510d16725a4d1dc031cb759dbc90bab427b0f60ddc6764681924dda848824605a4f08b7f526fe6bd4572458c94e83fbf2150f2eeb28d3011ec921996dc3e69efa52d5fcf3182b20b56b5857a926aa66605808079b4d52c0c0cfe06923fa92e65eeca2c3e6126108e8c1babf5ac522f4d7").unwrap();
@@ -432,7 +451,7 @@ async fn test_inclusion_proof_mocked() {
 #[ignore = "depends on external RPC"]
 #[tokio::test]
 async fn test_verify_batch() {
-    let cfg = EigenConfig::default();
+    let cfg = test_eigen_config();
     let query_client = create_remote_query_client();
     let verifier = Verifier::new(cfg, Arc::new(query_client)).await.unwrap();
     let cert = BlobInfo {
@@ -528,7 +547,7 @@ async fn test_verify_batch_mocked() {
     );
     mock_replies.insert(mock_req, mock_res);
 
-    let cfg = EigenConfig::default();
+    let cfg = test_eigen_config();
     let query_client = MockVerifierClient::new(mock_replies);
     let verifier = Verifier::new(cfg, Arc::new(query_client)).await.unwrap();
     let cert = BlobInfo {
@@ -608,7 +627,7 @@ async fn test_verify_batch_mocked() {
 #[ignore = "depends on external RPC"]
 #[tokio::test]
 async fn test_verify_security_params() {
-    let cfg = EigenConfig::default();
+    let cfg = test_eigen_config();
     let query_client = create_remote_query_client();
     let verifier = Verifier::new(cfg, Arc::new(query_client)).await.unwrap();
     let cert = BlobInfo {
@@ -717,7 +736,7 @@ async fn test_verify_securityt_params_mocked() {
     );
     mock_replies.insert(mock_req, mock_res);
 
-    let cfg = EigenConfig::default();
+    let cfg = test_eigen_config();
     let client = MockVerifierClient::new(mock_replies);
     let verifier = Verifier::new(cfg, Arc::new(client)).await.unwrap();
     let cert = BlobInfo {
